# backend/docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The API documentation for Panzerbot API endpoints.",
        "title": "Panzerbot API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/v1/peripheral/soundboard": {
            "get": {
                "description": "Returns the list of available soundboard track names",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "peripheral"
                ],
                "summary": "Get soundboard tracks",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/adaptersinhttpdto.PeripheralGetSoundboardTrackResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/peripheral/soundboard/stop": {
            "post": {
                "description": "Stops all currently playing soundboard tracks",
                "tags": [
                    "peripheral"
                ],
                "summary": "Stop all soundboard tracks",
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/peripheral/soundboard/{track_idx}": {
            "post": {
                "description": "Plays a soundboard track by its index",
                "tags": [
                    "peripheral"
                ],
                "summary": "Play a soundboard track",
                "parameters": [
                    {
                        "type": "integer",
                        "example": 0,
                        "description": "Track index",
                        "name": "track_idx",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/peripheral/ws": {
            "get": {
                "description": "Establish a WebSocket connection to control motor speed and servo angle in real-time",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "peripheral"
                ],
                "summary": "Peripheral WebSocket",
                "responses": {
                    "101": {
                        "description": "Switching Protocols"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/v1/rtc/ws": {
            "get": {
                "description": "Establish a WebSocket connection for WebRTC signalling. The robot sends an offer immediately after connecting. The peer replies with an answer and exchanges ICE candidates to establish the media connection. The robot streams camera and microphone; the peer may send audio back to the robot's speaker.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "rtc"
                ],
                "summary": "RTC Signalling WebSocket",
                "responses": {
                    "101": {
                        "description": "Switching Protocols"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/adaptersinhttpdto.CommonErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "adaptersinhttpdto.CommonErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "A specific error message"
                }
            }
        },
        "adaptersinhttpdto.PeripheralGetSoundboardTrackResponse": {
            "type": "object",
            "properties": {
                "tracks": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Boom",
                        "Fart",
                        "Get Out",
                        "Oi Oi Oi",
                        "Outro",
                        "Rizz",
                        "Uwu"
                    ]
                }
            }
        }
    }
}